% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{audit}
\alias{audit}
\title{Provide an R session audit}
\usage{
audit(filename = "", pbs_jobid = get_pbs_jobid())
}
\arguments{
\item{filename}{character, the name of the file to dump to or "" to print to console}

\item{pbs_jobid}{character, the PBS jobid if known}
}
\value{
NULL invisibly
}
\description{
Provide an R session audit
}
