% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlier-package.R
\docType{package}
\name{charlier-package}
\alias{charlier-package}
\alias{charlier}
\title{charlier-package}
\description{
Provides added tools for working within projects in the queued/charlie environment.
}
\section{Guide to functions}{

See \href{https://github.com/BigelowLab/charlier}{charlier} README for examples.
}

