% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{error}
\alias{error}
\title{A wrapper around the logging error level messaging}
\usage{
error(...)
}
\arguments{
\item{...}{arguments for \code{\link[futile.logger]{flog.error}}}
}
\description{
We rely on \href{https://CRAN.R-project.org/package=futile.logger}{futile.logger}, but that is subject to change.
}
