% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_extension}
\alias{get_extension}
\title{Retrieve file extensions}
\usage{
get_extension(
  filename = c("BR2_2016_S216_L001_R2_001.fastq", "foobar.fastq.gz", "fuzzbaz"),
  segments = 1
)
}
\arguments{
\item{filename}{character, vector of one or more filenames}

\item{segments}{integer, by default the last (1) segement, but use this to 
to retrieve the last n segments (like you want the 'tar.gz' not just 'gz')}
}
\value{
character vector of the last \code{segments} of the filename(s). If no 
  extension then NA_character_ is returned.
}
\description{
Retrieve file extensions
}
