% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_pbs_jobid}
\alias{get_pbs_jobid}
\title{Retrieve the PBS JOBID if available}
\usage{
get_pbs_jobid(no_pbs_text = "")
}
\arguments{
\item{no_pbs_text}{character, the text to return if no jobid found}
}
\value{
character jobid or missing_text value
}
\description{
Retrieve the PBS JOBID if available
}
