% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{info}
\alias{info}
\title{A wrapper around the logging informational level messaging.}
\usage{
info(...)
}
\arguments{
\item{...}{arguments for \code{\link[futile.logger]{flog.info}}}
}
\description{
We rely on \href{https://CRAN.R-project.org/package=futile.logger}{futile.logger}, but that is subject to change.
}
