% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.nullna}
\alias{is.nullna}
\title{Tests that an object is NULL nor NA}
\usage{
is.nullna(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical, TRUE if \code{x} is NULL or NA
}
\description{
\code{NULL} always returns TRUE.  Multiple element non-\code{NULL} objects always return FALSE.  If you are not
sure if your object might be length > 1 then use as \code{is.nullna(x[1])}.
}
