% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{munion}
\alias{munion}
\title{Compute the union of a list of vectors}
\usage{
munion(x)
}
\arguments{
\item{x}{a list of zero or more vectors}
}
\value{
a vector of the union or NULL is the input is empty
}
\description{
Compute the union of a list of vectors
}
\examples{
\dontrun{
  munion(list(x,y,z))
  # [1]  1  5  7  9  3  8 11 42  0  6  
} 
}
\seealso{
\code{\link[base]{union}}
}
