% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{package_imports}
\alias{package_imports}
\title{Retrieve a character vector of the packages impoerted by a specified package}
\usage{
package_imports(
  x = "charlier",
  collapse = FALSE,
  exclude = c("stats", "graphics", "grDevices", "utils", "datasets", "methods", "base")
)
}
\arguments{
\item{x}{character, one or more packages}

\item{collapse}{logical, if TRUE then collapse the packages into a single string}

\item{exclude}{character, vector of packages to exlude (base packages generally) or NULL to skip}
}
\value{
a named list of character package names or a single vector
}
\description{
Retrieve a character vector of the packages impoerted by a specified package
}
