% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read a configuration file}
\usage{
read_config(filename, autopopulate = FALSE, example = FALSE, ...)
}
\arguments{
\item{filename}{character, the full path and name of the config file to read}

\item{autopopulate}{logical, if TRUE then call \code{autopopulate_config}}

\item{example}{logical, if TRUE then ignore the filename and read the example configuration}

\item{...}{other arguments for \code{autopopulate_config}}
}
\value{
a named list of configuration elements or a character vector
}
\description{
Read a configuration file
}
