% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{sendmail}
\alias{sendmail}
\title{Send a mail message where \code{mail} binary is available}
\usage{
sendmail(
  to = "btupper@bigelow.org",
  subject = "mail from charlier",
  message = "It's so wonderful to get mail, ain't it?",
  attachment = NULL
)
}
\arguments{
\item{to}{character, vector of one or more email addresses}

\item{subject}{character subject line}

\item{message}{character, vector of one or more lines of text}

\item{attachment}{NULL or charcater, file path to optional attachment}
}
\value{
integer success code (0) or some non-zero if an error
}
\description{
Send a mail message where \code{mail} binary is available
}
\examples{
\dontrun{
cat(LETTERS, sep = "\n", file = "letters.txt") 
sendmail(to = "foo@bar.com", 
         subject = "Facts", 
         message = c("Fact: dogs start out as puppies.", "Not sure about cats."),
         attachment = "letters.txt")
}
}
