% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{start_logger}
\alias{start_logger}
\title{Start a \code{\link[futile.logger]{flog.logger}}}
\usage{
start_logger(filename = "./log", level = "info")
}
\arguments{
\item{filename}{character, NA or NULL  If not NULL or NA, then the name of the 
loging file to write to in addition to console logging}

\item{level}{character, by default "info" but complete options include 
trace, debug, info, warn, error, fatal  Setting the level to info allows for 
messages \code{flog.info(...)} through \code{flog.fatal(...)} to be printed.
Set it to lower values just for debugging.  Not that if the level is "info" then
you can still put a \code{flog.debug("blah")} in your script - it just won't print.}
}
\value{
NULL invisibly
}
\description{
We rely on \href{https://CRAN.R-project.org/package=futile.logger}{futile.logger}, but that is subject to change.
}
\details{
This doesn't actually start the logger, but it might feel like it.  Here you can simply
set an optional filename to echo the output to, and you can set the logging level. There
are a lot of things you can control about logging, and using this function doesn't prevent
you from using all of those other features.
}
\examples{
 \dontrun{
   library(futile.logger)
   start_logger(filename = "~/my_log_file")
   who <- "World"
   flog.info("Hello, \%s!", who)
 } 
}
