% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{strip_extension}
\alias{strip_extension}
\title{Strip the extension(s) off of one or more filenames}
\usage{
strip_extension(
  filename = c("BR2_2016_S216_L001_R2_001.fastq", "foobar.fastq.gz", "fuzzbaz.txt"),
  segments = 1,
  pattern = c(".fastq", NA)[2]
)
}
\arguments{
\item{filename}{character one or more filenames}

\item{segments}{integer, by default 1, ignored if pattern is not NA}

\item{pattern}{character, a pattern to strip. If not NA then 
\code{segments} is ignored.}
}
\value{
filename with extension stripped
}
\description{
Note if ext is ".fastq" then ".fastq.gz" and ".fastq.tar.gz" will also be stripped
}
