% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{suffix_filename}
\alias{suffix_filename}
\title{Modify a filename to add a suffix before the extension}
\usage{
suffix_filename(filename = "/this/is/foo.yaml", prefix = "-bar", segments = 1)
}
\arguments{
\item{filename}{character, vector of one or more filenames}

\item{prefix}{character, the prefix to insert before the file extension}

\item{segments}{integer, by default the last (1) segement, but use this to 
to retrieve the last n segments (like you want the 'tar.gz' not just 'gz')}
}
\description{
Modify a filename to add a suffix before the extension
}
