% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{warn}
\alias{warn}
\title{A wrapper around the logging warning level messaging}
\usage{
warn(...)
}
\arguments{
\item{...}{arguments for \code{\link[futile.logger]{flog.warn}}}
}
\description{
We rely on \href{https://CRAN.R-project.org/package=futile.logger}{futile.logger}, but that is subject to change.
}
