% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_config}
\alias{write_config}
\title{Write a configuration file}
\usage{
write_config(x, filename, ...)
}
\arguments{
\item{x}{a named list - ala yaml}

\item{filename}{character, the full path and filename to write to}

\item{...}{other arguments for \code{\link[yaml]{write_yaml}}}
}
\value{
the input named list, \code{x}
}
\description{
Write a configuration file
}
